chore(nodejs)!: deprecate node 12.x, 14.x and 16.x
BREAKING CHANGE: Now we will no longer support old nodejs versions
