+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs lts-gallium
+nodejs lts
